gemini-qwen-distill/requirements.txt
torch
transformers
trl
accelerate
datasets
python-dotenv
google-generativeai
